@ setPassiveClient 접속 종료 시간 지연 문제 버그 수정   - 로깅 기록에 몇만라인 단위로 로그가 기록되었음.

+++ src/device-controller/server/SocketClient.js
@@ -19,6 +19,9 @@ class SocketClient extends AbstController {
 
     this.configInfo = siteUUID;
 
+    // Socket Destroy 중인 상태
+    this.isDestroying = false;
+
     const foundInstance = _.find(instanceList, instanceInfo =>
       _.isEqual(instanceInfo.id, this.configInfo),
     );
@@ -56,14 +59,21 @@ class SocketClient extends AbstController {
    */
   setPassiveClient(client) {
     BU.CLI('setPassiveClient', this.configInfo);
+
+    BU.CLI(this.isDestroying);
+    if (this.isDestroying) {
+      BU.CLI('기존 접속 종료 중...');
+      return false;
+    }
     // 기본 Client와의 연결이 존재한다면 해당 연결을 끊음
     if (!_.isEmpty(this.client)) {
       BU.CLI('기존 연결 제거');
+      this.isDestroying = true;
       this.client.destroy();
-
-      // setTimeout(() => {
-      this.setPassiveClient(client);
-      // }, 1);
+      setTimeout(() => {
+        this.isDestroying = false;
+        this.setPassiveClient(client);
+      }, 1);
     } else {
       client.on('data', bufferData => {
         this.notifyData(bufferData);
